clarify the datalog engine
