+++ datalog/datalog.go
@@ -188,8 +188,6 @@ type Rule struct {
 	Head        Predicate
 	Body        []Predicate
 	Expressions []Expression
-
-	forbiddenIDs []Term
 }
 
 type InvalidRuleError struct {
@@ -204,9 +202,9 @@ func (e InvalidRuleError) Error() string {
 func (r Rule) Apply(facts *FactSet, newFacts *FactSet, syms *SymbolTable) error {
 	// extract all variables from the rule body
 	variables := make(MatchedVariables)
-	for _, p := range r.Body {
-		for _, id := range p.Terms {
-			v, ok := id.(Variable)
+	for _, predicate := range r.Body {
+		for _, term := range predicate.Terms {
+			v, ok := term.(Variable)
 			if !ok {
 				continue
 			}
@@ -214,33 +212,26 @@ func (r Rule) Apply(facts *FactSet, newFacts *FactSet, syms *SymbolTable) error 
 		}
 	}
 
-	combined, err := NewCombinator(variables, r.Body, r.Expressions, facts).Combine(syms)
+	combinations, err := NewCombinator(variables, r.Body, r.Expressions, facts).Combine(syms)
 	if err != nil {
 		return err
 	}
-outer:
-	for _, h := range combined {
-		p := r.Head.Clone()
-		for i, id := range p.Terms {
-			k, ok := id.(Variable)
+
+	for _, combined_variables := range combinations {
+		predicate := r.Head.Clone()
+		for i, term := range predicate.Terms {
+			k, ok := term.(Variable)
 			if !ok {
 				continue
 			}
-			v, ok := h[k]
+			v, ok := combined_variables[k]
 			if !ok {
 				return InvalidRuleError{r, k}
 			}
 
-			// prevent the rule from generating facts with forbidden IDs
-			for _, f := range r.forbiddenIDs {
-				if f.Equal(*v) {
-					continue outer
-				}
-			}
-
-			p.Terms[i] = *v
+			predicate.Terms[i] = *v
 		}
-		newFacts.Insert(Fact{p})
+		newFacts.Insert(Fact{predicate})
 	}
 
 	return nil
@@ -428,10 +419,6 @@ func (w *World) Query(pred Predicate) *FactSet {
 		if len(f.Predicate.Terms) != len(pred.Terms) {
 			continue
 		}
-		/*minLen := len(f.Predicate.IDs)
-		if l := len(pred.IDs); l < minLen {
-			minLen = l
-		}*/
 
 		matches := true
 		for i := 0; i < len(pred.Terms); i++ {
@@ -514,9 +501,9 @@ func NewCombinator(variables MatchedVariables, predicates []Predicate, expressio
 		allFacts:    allFacts,
 	}
 	currentFacts := make(FactSet, 0, len(*allFacts))
-	for _, f := range *allFacts {
-		if len(predicates) > 0 && f.Match(predicates[0]) {
-			currentFacts = append(currentFacts, f)
+	for _, fact := range *allFacts {
+		if len(predicates) > 0 && fact.Match(predicates[0]) {
+			currentFacts = append(currentFacts, fact)
 		}
 	}
 	c.currentFacts = &currentFacts
@@ -613,253 +600,4 @@ func (c *Combinator) Combine(syms *SymbolTable) ([]map[Variable]*Term, error) {
 		}
 	}
 	return variables, nil
-}
-
-var DEFAULT_SYMBOLS = [...]string{
-	"read",
-	"write",
-	"resource",
-	"operation",
-	"right",
-	"time",
-	"role",
-	"owner",
-	"tenant",
-	"namespace",
-	"user",
-	"team",
-	"service",
-	"admin",
-	"email",
-	"group",
-	"member",
-	"ip_address",
-	"client",
-	"client_ip",
-	"domain",
-	"path",
-	"version",
-	"cluster",
-	"node",
-	"hostname",
-	"nonce",
-	"query",
-}
-
-var OFFSET = 1024
-
-type SymbolTable []string
-
-func (t *SymbolTable) Insert(s string) String {
-	for i, v := range DEFAULT_SYMBOLS {
-		if string(v) == s {
-			return String(i)
-		}
-	}
-
-	for i, v := range *t {
-		if string(v) == s {
-			return String(OFFSET + i)
-		}
-	}
-	*t = append(*t, s)
-
-	return String(OFFSET + len(*t) - 1)
-}
-
-func (t *SymbolTable) Sym(s string) Term {
-	for i, v := range DEFAULT_SYMBOLS {
-		if string(v) == s {
-			return String(i)
-		}
-	}
-
-	for i, v := range *t {
-		if string(v) == s {
-			return String(OFFSET + i)
-		}
-	}
-	return nil
-}
-
-func (t *SymbolTable) Index(s string) uint64 {
-	for i, v := range DEFAULT_SYMBOLS {
-		if string(v) == s {
-			return uint64(i)
-		}
-	}
-
-	for i, v := range *t {
-		if string(v) == s {
-			return uint64(OFFSET + i)
-		}
-	}
-	panic("index not found")
-}
-
-func (t *SymbolTable) Str(sym String) string {
-	if int(sym) < 1024 {
-		if int(sym) > len(DEFAULT_SYMBOLS)-1 {
-			return fmt.Sprintf("<invalid symbol %d>", sym)
-		} else {
-			return DEFAULT_SYMBOLS[int(sym)]
-		}
-	}
-	if int(sym)-1024 > len(*t)-1 {
-		return fmt.Sprintf("<invalid symbol %d>", sym)
-	}
-	return (*t)[int(sym)-1024]
-}
-
-func (t *SymbolTable) Var(v Variable) string {
-	if int(v) < 1024 {
-		if int(v) > len(DEFAULT_SYMBOLS)-1 {
-			return fmt.Sprintf("<invalid variable %d>", v)
-		} else {
-			return DEFAULT_SYMBOLS[int(v)]
-		}
-	}
-	if int(v)-1024 > len(*t)-1 {
-		return fmt.Sprintf("<invalid variable %d>", v)
-	}
-	return (*t)[int(v)-1024]
-}
-
-func (t *SymbolTable) Clone() *SymbolTable {
-	newTable := *t
-	return &newTable
-}
-
-// SplitOff returns a newly allocated slice containing the elements in the range
-// [at, len). After the call, the receiver will be left containing
-// the elements [0, at) with its previous capacity unchanged.
-func (t *SymbolTable) SplitOff(at int) *SymbolTable {
-	if at > len(*t) {
-		panic("split index out of bound")
-	}
-
-	new := make(SymbolTable, len(*t)-at)
-	copy(new, (*t)[at:])
-
-	*t = (*t)[:at]
-
-	return &new
-}
-
-func (t *SymbolTable) Len() int {
-	return len(*t)
-}
-
-// IsDisjoint returns true if receiver has no elements in common with other.
-// This is equivalent to checking for an empty intersection.
-func (t *SymbolTable) IsDisjoint(other *SymbolTable) bool {
-	m := make(map[string]struct{}, len(*t))
-	for _, s := range *t {
-		m[s] = struct{}{}
-	}
-
-	for _, os := range *other {
-		if _, ok := m[os]; ok {
-			return false
-		}
-	}
-
-	return true
-}
-
-// Extend insert symbols from the given SymbolTable in the receiving one
-// excluding any Symbols already existing
-func (t *SymbolTable) Extend(other *SymbolTable) {
-	for _, s := range *other {
-		t.Insert(s)
-	}
-}
-
-type SymbolDebugger struct {
-	*SymbolTable
-}
-
-func (d SymbolDebugger) Predicate(p Predicate) string {
-	strs := make([]string, len(p.Terms))
-	for i, id := range p.Terms {
-		var s string
-		if sym, ok := id.(String); ok {
-			s = "\"" + d.Str(sym) + "\""
-		} else if variable, ok := id.(Variable); ok {
-			s = "$" + d.Var(variable)
-		} else {
-			s = fmt.Sprintf("%v", id)
-		}
-		strs[i] = s
-	}
-	return fmt.Sprintf("%s(%s)", d.Str(p.Name), strings.Join(strs, ", "))
-}
-
-func (d SymbolDebugger) Rule(r Rule) string {
-	head := d.Predicate(r.Head)
-	preds := make([]string, len(r.Body))
-	for i, p := range r.Body {
-		preds[i] = d.Predicate(p)
-	}
-	expressions := make([]string, len(r.Expressions))
-	for i, e := range r.Expressions {
-		expressions[i] = d.Expression(e)
-	}
-
-	var expressionsStart string
-	if len(preds) > 0 && len(expressions) > 0 {
-		expressionsStart = ", "
-	}
-
-	return fmt.Sprintf("%s <- %s%s%s", head, strings.Join(preds, ", "), expressionsStart, strings.Join(expressions, ", "))
-}
-
-func (d SymbolDebugger) CheckQuery(r Rule) string {
-	preds := make([]string, len(r.Body))
-	for i, p := range r.Body {
-		preds[i] = d.Predicate(p)
-	}
-	expressions := make([]string, len(r.Expressions))
-	for i, e := range r.Expressions {
-		expressions[i] = d.Expression(e)
-	}
-
-	var expressionsStart string
-	if len(preds) > 0 && len(expressions) > 0 {
-		expressionsStart = ", "
-	}
-
-	return fmt.Sprintf("%s%s%s", strings.Join(preds, ", "), expressionsStart, strings.Join(expressions, ", "))
-}
-
-func (d SymbolDebugger) Expression(e Expression) string {
-	return e.Print(d.SymbolTable)
-}
-
-func (d SymbolDebugger) Check(c Check) string {
-	queries := make([]string, len(c.Queries))
-	for i, q := range c.Queries {
-		queries[i] = d.CheckQuery(q)
-	}
-	return fmt.Sprintf("check if %s", strings.Join(queries, " or "))
-}
-
-func (d SymbolDebugger) World(w *World) string {
-	facts := make([]string, len(*w.facts))
-	for i, f := range *w.facts {
-		facts[i] = d.Predicate(f.Predicate)
-	}
-	rules := make([]string, len(w.rules))
-	for i, r := range w.rules {
-		rules[i] = d.Rule(r)
-	}
-	return fmt.Sprintf("World {{\n\tfacts: %v\n\trules: %v\n}}", facts, rules)
-}
-
-func (d SymbolDebugger) FactSet(s *FactSet) string {
-	strs := make([]string, len(*s))
-	for i, f := range *s {
-		strs[i] = d.Predicate(f.Predicate)
-	}
-	return fmt.Sprintf("%v", strs)
 }
